Touches up on the "Mix" button where it will display a confirmation panel
When a player would finish the drink and press the "Mix" button, a confirmation panel will display allowing the player to either serve or toss the drink. Serving the drink will complete the process and branch off into different scenarios depending if the character likes or dislikes the drink. Tossing the drink will close the panel and reset the ingredient values to zero so that the player can restart the drink.